requirements.txt
Flask==2.3.2
mss==9.0.1
Pillow==9.4.0
Pillow==11.0.0
pywin32==301
